make tablemixin exportable
